Split vm64 from vm using symlinks
Revert vm/ to master (32-bit) state and create vm64/ for the 64-bit
port. Unmodified files in vm64/ are symlinks back to vm/, while the
~179 modified/added files are real copies. Four entirely unchanged
directories (fast_compiler, cocoa, ppc, sparc) use directory-level
symlinks. Build targets are Self (32-bit) and Self64 (64-bit).

+++ vm/src/i386/fast_compiler/registerState_i386.cpp
@@ -40,7 +40,7 @@ Location RegisterState::pickLocal() {
 
 void RegisterState::allocateArgs(fint nargs, bool isPrimCall) {
   Unused(isPrimCall);
-  argDepth = max(argDepth, (int32)nargs);
+  argDepth = max(argDepth, nargs);
   rcvrDepth = 1;
 }
